refactor(explain): move Query Doctor out of setup() into its own module
detect_adapter, _parse_explain_nodes, TRANSLATIONS and _render_query_doctor
were declared inside setup(): ~200 lines rebuilt on every setup() call and
invisible to tests until the first one (explain_spec only passed because an
earlier spec in the run happened to call setup()). They are pure — no
upvalue from setup()'s scope, opts included — so they move to
lua/dadbod-grip/explain.lua unchanged.

M._parse_explain_nodes / M._render_query_doctor stay on the plugin module,
now bound at module level, so both the :GripExplain command and the existing
specs keep the names they used. setup() drops from ~900 to ~700 lines.

Verified by snapshotting parse_nodes + render (nodes, lines, highlight marks)
over postgresql/mysql/duckdb/sqlite/unrecognised/empty EXPLAIN fixtures
before and after: identical. :GripExplain end-to-end on tests/seed_sqlite.db
still opens the float with the same content and 3 highlight extmarks.
